LPS-66061 move and export classes or else LangBuilder "gw buildLang" doesn't actually work
